+++ floris/tools/__init__.py
@@ -39,7 +39,6 @@ Examples:
 from .floris_interface import FlorisInterface
 from .floris_interface_legacy_reader import FlorisInterfaceLegacyV2
 from .parallel_computing_interface import ParallelComputingInterface
-from .turbine_utilities import build_cosine_loss_turbine_dict, check_smooth_power_curve
 from .uncertainty_interface import UncertaintyInterface
 from .visualization import (
     plot_rotor_values,
